Update belt_control.ipynb
When running the notebook, it didn't work for me because I was copying what looked like the UUID in the last column.  As it turns out, the id to type in is the first data column (after the counter), which in my case was all BTLE MAC addresses, not the long UUIDs from the example.  This causes confusion and is probably outdated?